+++ coloraide/color.py
@@ -56,10 +56,11 @@ from .filters import Filter
 from .filters.w3c_filter_effects import Sepia, Brightness, Contrast, Saturate, Opacity, HueRotate, Grayscale, Invert
 from .filters.cvd import Protan, Deutan, Tritan
 from .interpolate import Interpolator, Interpolate
+from .interpolate.linear import Linear
+from .interpolate.continuous import Continuous
 from .interpolate.bspline import BSpline
 from .interpolate.bspline_natural import NaturalBSpline
 from .interpolate.monotone import Monotone
-from .interpolate.linear import Linear
 from .types import Plugin
 from typing import overload, Sequence, Any, Callable, Mapping
 
@@ -1096,6 +1097,7 @@ Color.register(
 
         # Interpolation
         Linear(),
+        Continuous(),
         BSpline(),
         NaturalBSpline(),
         Monotone()